+++ providers/sqlite3/provider.go
@@ -14,8 +14,8 @@ var initQueries = []string{
 	`CREATE TABLE IF NOT EXISTS %s (
 		id VARCHAR(64) PRIMARY KEY NOT NULL DEFAULT '',
 		data TEXT NOT NULL,
-		last_active INT(10) NOT NULL DEFAULT '0',
-		expiration INT(10) NOT NULL DEFAULT '0'
+		last_active BIGINT NOT NULL DEFAULT '0',
+		expiration INT NOT NULL DEFAULT '0'
 	);`,
 	"CREATE INDEX last_active ON %s (last_active);",
 	"CREATE INDEX expiration ON %s (expiration);",
